Controller no longer requires a com port to be connected. If you enter "-1" as the port to use (or in fact, any port lower than 0) then it won't even try to connect.  -1 is the default.  This is because some machines do not have a com port at all.
I have also reordered the bit of code that watches for mouse clicks, and given priority to the panel.  This means that if you change your machine size so that it overlaps with the panel, you can still use the panel buttons.  Previously, although the panel was visually on top of the machine preview, all of the clicks hit the machine rather than the panel.

What else.  Have made the status of the machine a bit more obvious with some bonny colours.  Same with the queue header.  

git-svn-id: http://polargraph.googlecode.com/svn/controller/trunk@108 5ebda15a-75ab-8a5e-e99a-b4cc1c219dce